Add additional dynamic data options to Page Title
